requirements.txt
# VibeManga Requirements
# Python 3.8+

# Core dependencies
click>=8.0.0
python-dotenv>=0.19.0
rich>=13.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Optional: CBR file support
rarfile>=4.0

# Notes:
# - Keep versions conservative; adjust if you have stricter constraints.
# VibeManga Requirements
# Python 3.8+

# Core dependencies
click>=8.0.0
python-dotenv>=0.19.0
rich>=13.0.0

# Optional: CBR file support
rarfile>=4.0
 